move settings import tests to a separate file
